feat: Add branch age calculation script
A new Python script has been added to calculate the age of each release branch in a Git repository. The script fetches all branches, filters for release branches, and then determines the oldest commit in the main branch that's not in each release branch. It calculates the age of each branch by comparing the date of this 'fork' commit with the date of the latest commit on that branch. The results are sorted by latest commit date and printed out.